[OBMIN-353] Moving the NavCategory component
