+++ src/components/category-button/category-nav/config.ts
@@ -1,4 +1,4 @@
-import * as Images from '../img';
+import * as Images from '../../img';
 
 export const images = [
   { img: Images.clothes, sun: Images.sunClothes },
